# Conflicts:
#	bunhobono/frontend/src/features/member/MemSignup.vue

+++ bunhobono/backend/src/main/java/api/a_security_config/LoginDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class LoginDTO {
-    private long memberNo;
+    private Integer memberNo;
     private String loginId,loginPwd,role,memStatus;
 }
